Separate keyed browser tab rendering from panel lifecycle
