+++ src/Ai.Tlbx.MidTerm/src/ts/modules/web/webPanel.ts
@@ -5,7 +5,6 @@
  */
 
 import { $webPreviewUrl, $activeSessionId } from '../../stores';
-import { reconcileKeyedChildren } from '../../utils/domReconcile';
 import { applyStoredViewportToFrame } from './webViewport';
 import {
   clearWebPreviewState,
@@ -37,7 +36,7 @@ import {
   stripInternalPreviewQueryParams,
 } from './previewProxyUrl';
 import { openTopLevelPreview, resolveTopLevelProxyUrl } from './webTopLevelHandoff';
-import { buildPreviewTabLabel, shouldRenderPreviewTab } from './webPreviewTabLabel';
+import { renderBrowserPreviewTabs } from './webPreviewTabs';
 import {
   getActiveDockedClient,
   getActivePreview,
@@ -45,7 +44,6 @@ import {
   getActiveUrl,
   getSessionDockedClient,
   getSessionPreview,
-  listSessionPreviews,
   setActiveMode,
   setActiveUrl,
   setSessionDockedClient,
@@ -121,76 +119,12 @@ export function renderPreviewTabs(): void {
 
   const sessionId = $activeSessionId.get();
   const selectedPreviewName = getActivePreviewName();
-  const previews = sessionId ? listSessionPreviews(sessionId) : [];
-  previewTabs.setAttribute('role', 'tablist');
-  previewTabs.setAttribute('aria-label', 'Browser previews');
-  reconcileKeyedChildren(
+  renderBrowserPreviewTabs(
     previewTabs,
-    previews.filter((preview) =>
-      shouldRenderPreviewTab(preview, selectedPreviewName, previews.length),
-    ),
-    {
-      key: (preview) => `${sessionId}:${preview.previewName}`,
-      create: (preview) => {
-        const tab = document.createElement('div');
-        tab.className = 'web-preview-tab-shell';
-        tab.dataset.previewName = preview.previewName;
-        const button = document.createElement('button');
-        button.type = 'button';
-        button.className = 'web-preview-tab';
-        button.setAttribute('role', 'tab');
-        button.addEventListener('click', () => {
-          previewTabSelectHandler?.(preview.previewName);
-        });
-        button.addEventListener('keydown', (event) => {
-          const buttons = Array.from(
-            previewTabs?.querySelectorAll<HTMLButtonElement>('[role="tab"]') ?? [],
-          );
-          const index = buttons.indexOf(button);
-          let next: HTMLButtonElement | undefined;
-          if (event.key === 'ArrowRight') next = buttons[(index + 1) % buttons.length];
-          if (event.key === 'ArrowLeft')
-            next = buttons[(index + buttons.length - 1) % buttons.length];
-          if (event.key === 'Home') next = buttons[0];
-          if (event.key === 'End') next = buttons[buttons.length - 1];
-          if (!next) return;
-          event.preventDefault();
-          next.click();
-          next.focus();
-          next.scrollIntoView({ block: 'nearest', inline: 'nearest' });
-        });
-        tab.appendChild(button);
-
-        const closeButton = document.createElement('button');
-        closeButton.type = 'button';
-        closeButton.className = 'web-preview-tab-close';
-        closeButton.textContent = '×';
-        closeButton.addEventListener('click', (event) => {
-          event.preventDefault();
-          event.stopPropagation();
-          previewTabCloseHandler?.(preview.previewName);
-        });
-        tab.appendChild(closeButton);
-
-        return tab;
-      },
-      patch: (tab, preview) => {
-        const selected = preview.previewName === selectedPreviewName;
-        tab.classList.toggle('active', selected);
-        tab.classList.toggle('detached', preview.mode === 'detached');
-        tab.classList.toggle('empty', !preview.url);
-        const label = buildPreviewTabLabel(preview.url);
-        const button = tab.querySelector<HTMLButtonElement>('.web-preview-tab')!;
-        if (button.textContent !== label) button.textContent = label;
-        button.title = preview.url?.trim() || label;
-        button.setAttribute('aria-label', `Preview tab ${label}`);
-        button.setAttribute('aria-selected', String(selected));
-        button.tabIndex = selected ? 0 : -1;
-        const close = tab.querySelector<HTMLButtonElement>('.web-preview-tab-close')!;
-        close.title = `Close ${label}`;
-        close.setAttribute('aria-label', `Close preview tab ${label}`);
-      },
-    },
+    sessionId,
+    selectedPreviewName,
+    (name) => previewTabSelectHandler?.(name),
+    (name) => previewTabCloseHandler?.(name),
   );
 
   updateScreenshotButtonState();
